Added javadocs and removed tests that don't add value to the demo
