+++ chaos-monkey-spring-boot/pom.xml
@@ -123,7 +123,6 @@
         <dependency>
             <groupId>no.finn.unleash</groupId>
             <artifactId>unleash-client-java</artifactId>
-            <version>4.2.1</version>
             <optional>true</optional>
         </dependency>
