+++ chaos-monkey-spring-boot/src/test/java/de/codecentric/spring/boot/chaos/monkey/configuration/toggles/DefaultChaosTogglesTest.java
@@ -15,7 +15,7 @@ class DefaultChaosTogglesTest {
   }
 
   @Test
-  public void defaultToggleShouldBeEnabledAlwaysl() {
+  public void defaultToggleShouldBeEnabledAlways() {
     assertTrue(sut.isEnabled("chaos.monkey.repository"));
     assertTrue(sut.isEnabled(""));
   }
