+++ demo-apps/chaos-monkey-demo-app-unleash-toggles/src/main/java/com/example/chaos/monkey/toggledemo/UserAwareFakeUnleash.java
@@ -18,7 +18,6 @@ public class UserAwareFakeUnleash implements Unleash {
   private final UnleashContextProvider contextProvider;
 
   public UserAwareFakeUnleash(UnleashContextProvider contextProvider) {
-
     this.contextProvider = contextProvider;
   }
