+++ demo-apps/chaos-monkey-demo-app-unleash-toggles/src/main/java/com/example/chaos/monkey/toggledemo/WebSecurityConfig.java
@@ -1,0 +1,55 @@
+package com.example.chaos.monkey.toggledemo;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.provisioning.InMemoryUserDetailsManager;
+
+@Configuration
+@EnableWebSecurity
+public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
+
+  private final PasswordEncoder passwordEncoder;
+
+  public WebSecurityConfig(PasswordEncoder passwordEncoder) {
+    this.passwordEncoder = passwordEncoder;
+  }
+
+  @Override
+  protected void configure(HttpSecurity http) throws Exception {
+    http.authorizeRequests()
+        .antMatchers("/", "/home")
+        .permitAll()
+        .anyRequest()
+        .authenticated()
+        .and()
+        .httpBasic()
+        .and()
+        .logout()
+        .permitAll();
+  }
+
+  @Bean
+  @Override
+  public UserDetailsService userDetailsService() {
+    UserDetails user =
+        User.withUsername("user")
+            .password(passwordEncoder.encode("password"))
+            .roles("USER")
+            .build();
+
+    UserDetails chaosuser =
+        User.withUsername("chaosuser")
+            .password(passwordEncoder.encode("password"))
+            .roles("USER")
+            .build();
+
+    return new InMemoryUserDetailsManager(user, chaosuser);
+  }
+}
