--- demo-apps/chaos-monkey-demo-app-unleash-toggles/src/test/java/com/example/chaos/monkey/toggledemo/controller/HelloComponentIntegrationTest.java
@@ -1,54 +1,0 @@
-/*
- * Copyright 2020 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package com.example.chaos.monkey.toggledemo.controller;
-
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.core.IsNull.notNullValue;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import com.example.chaos.monkey.toggledemo.ChaosDemoApplication;
-import com.example.chaos.monkey.toggledemo.component.HelloComponent;
-import de.codecentric.spring.boot.chaos.monkey.configuration.ChaosMonkeyConfiguration;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.TestPropertySource;
-
-/** @author Benjamin Wilms */
-@SpringBootTest(
-    classes = ChaosDemoApplication.class,
-    webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@TestPropertySource("classpath:application-component-test.properties")
-public class HelloComponentIntegrationTest {
-
-  @Autowired private HelloComponent helloComponent;
-
-  @Autowired private ChaosMonkeyConfiguration chaosMonkeyConfiguration;
-
-  @Test
-  public void contextLoads() {
-    assertThat(helloComponent, notNullValue());
-  }
-
-  @Test
-  public void callingPublicMethodonComponent() {
-    assertTrue(chaosMonkeyConfiguration.settings().getWatcherProperties().isComponent());
-
-    helloComponent.sayHello();
-  }
-}
