--- demo-apps/chaos-monkey-demo-app-unleash-toggles/src/test/resources/application-component-test.properties
@@ -1,31 +1,0 @@
-#
-# Copyright 2018 the original author or authors.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-spring.profiles.active=chaos-monkey
-chaos.monkey.enabled=true
-chaos.monkey.assaults.latencyActive=true
-chaos.monkey.assaults.latencyRangeStart=1000
-chaos.monkey.assaults.latencyRangeEnd=4500
-chaos.monkey.watcher.controller=false
-chaos.monkey.watcher.component=true
-chaos.monkey.watcher.restController=false
-chaos.monkey.watcher.service=false
-chaos.monkey.watcher.repository=false
-chaos.monkey.assaults.level=1
-management.endpoint.chaosmonkey.enabled=true
-management.endpoint.chaosmonkeyjmx.enabled=false
-management.endpoints.web.exposure.include=*
-management.server.port=8888
-
